+++ cmd/geth/usage.go
@@ -59,9 +59,8 @@ var AppHelpFlagGroups = []flagGroup{
 			ChainIdentityFlag,
 			KeyStoreDirFlag,
 			NetworkIdFlag,
-			TestNetFlag,
 			DevModeFlag,
-			IdentityFlag,
+			NodeNameFlag,
 			FastSyncFlag,
 			LightKDFFlag,
 			CacheFlag,
@@ -138,6 +137,7 @@ var AppHelpFlagGroups = []flagGroup{
 		Name: "LOGGING AND DEBUGGING",
 		Flags: []cli.Flag{
 			VerbosityFlag,
+			LogDirFlag,
 			VModuleFlag,
 			BacktraceAtFlag,
 			MetricsFlag,
@@ -149,6 +149,13 @@ var AppHelpFlagGroups = []flagGroup{
 		Flags: []cli.Flag{
 			WhisperEnabledFlag,
 			NatspecEnabledFlag,
+		},
+	},
+	{
+		Name: "LEGACY",
+		Flags: []cli.Flag{
+			TestNetFlag,
+			Unused1,
 		},
 	},
 	{