+++ CHANGELOG.md
@@ -19,6 +19,17 @@ Releases considered stable may be found on our [Releases Page](https://github.co
 
 ## [Unreleased]
 
+#### Added
+- Newly configurable in external `chain.json`:
+    - `"state": { "startingNonce": NUMBER }` - _optional_ (mainnet: 0, morden: 1048576) - "dirty" starting world state
+    - `"network": NUMBER` - _required_ (mainnet: 1, morden: 2) - network id used to identify valid peers
+    - `"consensus": STRING` - _optional_ (default: "ethash", optional: "ethash-test") - specify smaller and faster pow algorithm, e.g. `--dev` mode sets "ethash-test"
+    > See cmd/geth/config/*.json for updated examples.
+
+- Dev mode (`--dev`) made compatible with `--chain` 
+
+#### Fixed
+- `geth attach` command uses chain subdirectory schema by default, e.g. `datadir/mainnet/geth.ipc` instead of `datadir/geth.ipc`
 
 ## [3.5.0] - 2017-06-02